Улучшение meta_desctiption: внедрение переменной {page} для предотвращения создания страниц с одинаковым Meta descripion. Исправлены мелкие ошибки.

+++ install/data_install/tables.php
@@ -1,4 +1,10 @@
 <?php
+/*
+	Copyright © Eleanor CMS, developed by Alexander Sunvas*, interface created by Rumin Sergey.
+	For details, visit the web site http://eleanor-cms.ru, emails send to support@eleanor-cms.ru .
+	*Pseudonym
+*/
+if(!defined('CMS'))die;
 $tables[]='SET FOREIGN_KEY_CHECKS=0;';
 
 $tables[]="DROP TABLE IF EXISTS `{$prefix}blocks`";
